requirements.txt
flask
flask-cors
pandas
openai
python-dotenv
